LPS-54637 migrate to module
